[test] Extract Button type test into separate file
